+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ## Release 1.3.39
 ### Changed
 - **Carousel**
-    - fix block padding and arrow position
+    - add new theme "landing"
 
 ## Release 1.3.38
 ### Changed